Fixed file count query in PR pass/fail evaluation
There are really 2 files counts - the number of files, and the total
number of predicted/observed tables in those files. We were previously
confusing the two numbers and comparing the accepted stats' number of
p/o tables vs the current pull request's number of files.

This raises the question of how this ever worked to begin with...