Fixed bug in missing files display in portal
Previously, we grabbed each p/o table from the accepted stats and each
p/o table from the current pull request. If any table name appeared
only once in the result, then we display an error message that that
table was missing from the current pull request.

The problem with this approach is that if the pull request adds a new
table, then there would only be one instance of that table name in the
result and thus the portal would display an error message saying that
this table is missing from the pull request.

I've fixed this so that the error message will only be shown if the
table exists in the accepted stats but not in the stats for the current
pull request. I've also added another message showing new table names
which were added by this pull request.