Errors while adding P/O values should be fatal
This should make Jenkins run red if the performance tests webservice
runs into problems.

There are currently a lot of missing records in the
PredictedObservedValues table. These missing records cause graphs to
fail to appear on certain pages on the portal. The missing data seems
to be due to random discrepancies in the tables in the DB and/or
related to bugs in this code. The errors are all written out to a log
file which no one checks. I'm hoping that making these errors fatal
will force the issue and allow the errors to be fixed.