Fixed a bug caused by using SimulationID to match predicted/observed data
Using SimulationID as the field name to match on causes pain because
one of the parameters we're about to pass into the stored proc is a
data table. The table must have a SimulationID column and it must have
a column for each of the field names to match on. We can't have two
columns called SimulationID, so we need to remap/rename some of the
columns when creating the table (selectedData) which gets passed into
the stored proc.