Fixed a bug where tests are sometimes not added to DB
The accepted value will always be null when someone adds a new test.
This results in the test never being added to the DB.

The real kicker is that the percentage passed calculation doesn't actually
consider what's actually in the DB, it works on all data even if the
accepted or current values are null.